Add overlay to oil fluid and barrel
This was tricky. Barrels are created in base/data-updates, so it is
impossible to modify them in our data.lua file because it does not exist
yet. It would be possible to do just barrels in a data-updates.lua, but
it is much simpler to move everything there. This also means any mods
that change icons in data.lua will be supported with overlaying them.