Replace heavy and light oil icons
The molecules were pretty confusing visually. These new icons are more
similar to vanilla, and don't lose the color for those who can use it.

Since they are so similar to vanilla, there is no point in overlaying
them over the base icon. So I removed that as an option.