Add custom technology background colors
The options are very limited, since it only pulls from the gui_new.png
file. The docs claim they can be passed a Sprite objects, so
theoretically they could be loaded from other files, but that is a lot
of color design work that I would rather skip for now.

Also update the senario to include turrets of all kinds to be able to
test the overlays from the previous commit. Also have some technologies
not researched to test the tech background colors.