Add localization for dropdown setting options
This was a huge pain. There is no place to define a localization string
that applies to all settings that use that "string ID". Each setting
needs its own values.

The "hack" is to generate the locale file at build time. Load the
settings file in a Lua script to get the list of all settings and their
possible string values, then create the locale file using the base
strings.

This will make it much easier for anyone to do a translation.

My Russian is super rusty, so I would be supprised if I didn't make a
lot of gender mistakes.