Improve compatibility with mods by prevent crashes
If other mods remove prototypes from base mod, our direct access of them
would cause a crash. (Ideally these mods would simply disable the
prototypes instead of deleting them, but clearly we can't rely on that.)

If a prototype we wanted to modify does not exist, print a warning to
the log and skip it.