Work around bug in Ultracube
This bug is pretty minor, and only happens when a mod modifies the icons
of a base prototype before Ultracube can copy it.

If this compat wasn't added, we would need to wait for
https://github.com/grandseiken/factorio-ultracube/pull/26
to be merged, and then lock our dependent version of Ultracube to that.
I don't see a good reason to be so restrictive, so:

Update the dependent version of Ultracube to 0.2.0 (the first version).
I tested it, and it works just fine.

This commit could be reverted if the PR is merged, but then we would
need to bump the dependent version, so I doubt I will do that.