Add icons to the visualization of agricultural tower
Wube should simply allow for customizing the visualization range of each
color separately, but until then, this is a hack that works.

When hovering an agricultural tower, check the 3x3 boxes and give them
an icon based on their color. This is done per player, so each player
gets some local storage to keep their render objects so we can delete
them later.

Since this is done at runtime, these can be runtime settings.

This has limitations though: since the API does not expose a way for
getting the cursor location, there is no way to apply these icons when
an agricultural tower is in the cursor. Similarly, there is no way to do
so when an agricultural tower is in the blueprint that is in the cursor.

Additionally, I made the logic of what "color" to apply the same that
the agricultural tower uses. This means it will show things as "green"
even when the tower is unable to plant there because different types of
soil are mixed (meaning no plant is eligible). Hopefully Wube fixes this
at the same time they fix the API limiations.