Downgrade map version of test palette scenario
While the mod claims to support Factorio version 2.0.7, the test palette
scenario was saved with a much newer version, which prevents loading it
in older versions. The map was recreated in version 2.0.7, which allows
it to be loaded in any version since then, and forces me to test the mod
in that version.

Also update a few things, like legacy to new rails.