Fix crash on missing planner icons
While not documented in the changelog, sometime around 2.0.34, the 32
pixel wide shortcut icons were removed and replaced with 56 pixel wide
versions. This caused a crash as the mod referenced the images directly.

Instead, copy the icons that the shortcut uses. Not only is this more
future proof, it allows the mod to run on both old and new versions.