Fix small module icons
Since Factorio 2.0.34, the engine is better about making sure composite
icons do not overflow their bounding box. This means it has been
resizing a few of these icons, mainly the modules.

In the same version they added a new "floating" flag, which disables
this bounding box check for that layer.

Rather than fight with the engine to keep it from resizing the icons, if
I manually check and it doesn't look like it is overflowing, then flag
it as floating.